requirements.txt
selenium
beautifulsoup4
lxml
pymongo
fastapi
uvicorn
pycryptodome